test/requirements.txt
micropython-json
micropython-time
micropython-os
micropython-uasyncio
